refactor: move mlx-nn into mlx-rs (#173)
* support specifying root in ModuleParameters macro

* move mlx-nn into mlx_rs

* updated imports

* fix import path

* clippy and fmt

* reorganize value_and_grad
