+++ Cargo.toml
@@ -10,7 +10,6 @@ members = [
     "mlx-macros", 
     "mlx-sys", 
     "mlx-rs", 
-    "mlx-nn",
     "mlx-internal-macros",
     "mlx-tests",
     "examples/*", 
@@ -30,3 +29,15 @@ mlx-nn = { version = "0.14.0", path = "mlx-nn" }
 thiserror = "1"
 float_eq = "1"
 pretty_assertions = "1.4.0"
+dyn-clone = "1"
+half = "2"
+mach-sys = "0.5"
+num-complex = "0.4"
+num_enum = "0.7"
+num-traits = "0.2"
+paste = "1"
+smallvec = "1"
+strum = { version = "0.26", features = ["derive"] }
+libc = "0.2"
+parking_lot = "0.12"
+tempfile = "3.13.0"
