+++ mlx-macros/src/lib.rs
@@ -1,6 +1,5 @@
 extern crate proc_macro;
 use proc_macro::TokenStream;
-use quote::quote;
 use syn::{parse_macro_input, DeriveInput};
 
 mod module_parameters;
@@ -42,16 +41,9 @@ mod module_parameters;
 ///     a: Param<Array>,
 /// }
 /// ```
-#[proc_macro_derive(ModuleParameters, attributes(param))]
+#[proc_macro_derive(ModuleParameters, attributes(module, param))]
 pub fn derive_module_parameters(input: TokenStream) -> TokenStream {
     let input = parse_macro_input!(input as DeriveInput);
     let module_param_impl = module_parameters::expand_module_parameters(&input).unwrap();
-
-    let output = quote! {
-        const _: () = {
-            extern crate mlx_rs as _mlx_rs;
-            #module_param_impl
-        };
-    };
-    TokenStream::from(output)
+    TokenStream::from(module_param_impl)
 }
