--- mlx-nn/Cargo.toml
@@ -1,16 +1,0 @@
-[package]
-name = "mlx-nn"
-version.workspace = true
-authors.workspace = true
-edition = "2021"
-
-[dependencies]
-mlx-internal-macros.workspace = true
-mlx-macros.workspace = true
-mlx-rs.workspace = true
-thiserror.workspace = true
-dyn-clone = "1"
-
-[dev-dependencies]
-float_eq.workspace = true
-pretty_assertions.workspace = true
