--- mlx-nn/src/macros.rs
@@ -1,3 +1,0 @@
-//! Macro re-exports.
-
-pub use mlx_macros::ModuleParameters;
