+++ mlx-rs/src/lib.rs
@@ -15,8 +15,10 @@ pub mod error;
 pub mod fast;
 pub mod fft;
 pub mod linalg;
+pub mod losses;
 pub mod module;
 pub mod nested;
+pub mod nn;
 pub mod ops;
 pub mod optimizers;
 pub mod random;
