+++ mlx-rs/src/losses.rs
@@ -1,7 +1,6 @@
 //! Loss functions
 
-use mlx_internal_macros::{generate_builder, Buildable};
-use mlx_rs::{
+use crate::{
     array,
     error::{CrossEntropyBuildError, Exception},
     ops::{
@@ -10,6 +9,7 @@ use mlx_rs::{
     },
     Array,
 };
+use mlx_internal_macros::{generate_builder, Buildable};
 
 #[inline]
 fn check_shape(
@@ -58,7 +58,9 @@ pub type CrossEntropyBuilderWeights<'a> = &'a Array;
 generate_builder! {
     /// Cross entropy loss function.
     #[derive(Debug, Clone, Buildable)]
+    #[buildable(root = crate)]
     #[builder(
+        root = crate,
         build_with = build_cross_entropy,
         err = CrossEntropyBuildError
     )]
@@ -170,6 +172,8 @@ generate_builder! {
     /// the loss calculation clips the input probabilities (in log-space) to a minimum value
     /// of `-100`.
     #[derive(Debug, Clone, Buildable)]
+    #[buildable(root = crate)]
+    #[builder(root = crate)]
     pub struct BinaryCrossEntropy<'a> {
         /// Optional weights for each target
         #[builder(optional, default = BinaryCrossEntropy::DEFAULT_WEIGHTS)]
@@ -237,6 +241,8 @@ impl<'a> BinaryCrossEntropy<'a> {
 generate_builder! {
     /// Computes the L1 loss
     #[derive(Debug, Clone, Buildable)]
+    #[buildable(root = crate)]
+    #[builder(root = crate)]
     pub struct L1Loss {
         /// Reduction type. Default to [`L1loss::DEFAULT_REDUCTION`]
         #[builder(optional, default = L1Loss::DEFAULT_REDUCTION)]
@@ -272,6 +278,8 @@ impl L1Loss {
 generate_builder! {
     /// Computes the mean squared error loss.
     #[derive(Debug, Clone, Buildable)]
+    #[buildable(root = crate)]
+    #[builder(root = crate)]
     pub struct MseLoss {
         /// Reduction type. Default to [`MseLoss::DEFAULT_REDUCTION`]
         #[builder(optional, default = MseLoss::DEFAULT_REDUCTION)]
@@ -307,6 +315,8 @@ impl MseLoss {
 generate_builder! {
     /// Computes the negative log likelihood loss.
     #[derive(Debug, Clone, Buildable)]
+    #[buildable(root = crate)]
+    #[builder(root = crate)]
     pub struct NllLoss {
         /// distribution axis. Default to [`NllLoss::DEFAULT_AXIS`]
         #[builder(optional, default = NllLoss::DEFAULT_AXIS)]
@@ -349,6 +359,8 @@ impl NllLoss {
 generate_builder! {
     /// Compute the negative log likelihood loss for a Gaussian distribution.
     #[derive(Debug, Clone, Buildable)]
+    #[buildable(root = crate)]
+    #[builder(root = crate)]
     pub struct GaussianNllLoss {
         /// Whether to include the constant term in the loss calculation. Default to
         /// [`GaussianNllLoss::DEFAULT_FULL`]
@@ -421,6 +433,8 @@ generate_builder! {
     /// sum(exp(targets) * (targets - inputs), axis, None)
     /// ```
     #[derive(Debug, Clone, Buildable)]
+    #[buildable(root = crate)]
+    #[builder(root = crate)]
     pub struct KlDivLoss {
         /// The distribution axis. Default to [`KlDivLoss::DEFAULT_AXIS`]
         #[builder(optional, default = KlDivLoss::DEFAULT_AXIS)]
@@ -471,6 +485,8 @@ generate_builder! {
     /// difference with a squared difference when the absolute difference is less
     /// than `beta`.
     #[derive(Debug, Clone, Buildable)]
+    #[buildable(root = crate)]
+    #[builder(root = crate)]
     pub struct SmoothL1Loss {
         /// The threshold after which the loss changes from the squared to the absolute difference.
         /// Default to [`SmoothL1Loss::DEFAULT_BETA`]
@@ -523,6 +539,8 @@ generate_builder! {
     /// Computes the triplet loss for a set of anchor, positive, and negative samples. Margin is
     /// represented with alpha in the math section.
     #[derive(Debug, Clone, Buildable)]
+    #[buildable(root = crate)]
+    #[builder(root = crate)]
     pub struct TripletLoss {
         /// Distribution axis. Default to [`TripletLoss::DEFAULT_AXIS`]
         #[builder(optional, default = TripletLoss::DEFAULT_AXIS)]
@@ -607,6 +625,8 @@ impl TripletLoss {
 generate_builder! {
     /// Compute the hinge loss.
     #[derive(Debug, Clone, Buildable)]
+    #[buildable(root = crate)]
+    #[builder(root = crate)]
     pub struct HingeLoss {
         /// Reduction type. Default to [`HingeLoss::DEFAULT_REDUCTION`]
         #[builder(optional, default = HingeLoss::DEFAULT_REDUCTION)]
@@ -643,6 +663,8 @@ impl HingeLoss {
 generate_builder! {
     /// Compute the Huber loss.
     #[derive(Debug, Clone, Buildable)]
+    #[buildable(root = crate)]
+    #[builder(root = crate)]
     pub struct HuberLoss {
         /// The threshold at which to change between L1 and L2 loss. Default to
         /// [`HuberLoss::DEFAULT_DELTA`]
@@ -696,6 +718,8 @@ generate_builder! {
     /// and like the L1 loss for large errors, reducing sensitivity to outliers. This
     /// dual behavior offers a balanced, robust approach for regression tasks.
     #[derive(Debug, Clone, Buildable)]
+    #[buildable(root = crate)]
+    #[builder(root = crate)]
     pub struct LogCoshLoss {
         /// Reduction type. Default to [`LogCoshLoss::DEFAULT_REDUCTION`]
         #[builder(optional, default = LogCoshLoss::DEFAULT_REDUCTION)]
@@ -732,6 +756,8 @@ impl LogCoshLoss {
 generate_builder! {
     /// Computes the cosine similarity loss.
     #[derive(Debug, Clone, Buildable)]
+    #[buildable(root = crate)]
+    #[builder(root = crate)]
     pub struct CosineSimilarityLoss {
         /// Embedding axis. Default to [`CosineSimilarityLoss::DEFAULT_AXIS`]
         #[builder(optional, default = CosineSimilarityLoss::DEFAULT_AXIS)]
@@ -793,6 +819,8 @@ impl CosineSimilarityLoss {
 generate_builder! {
     /// Computes the margin ranking loss.
     #[derive(Debug, Clone, Buildable)]
+    #[buildable(root = crate)]
+    #[builder(root = crate)]
     pub struct MarginRankingLoss {
         /// The margin by which the scores should be separated. Default to
         /// [`MarginRankingLoss::DEFAULT_MARGIN`]
@@ -848,8 +876,8 @@ impl MarginRankingLoss {
 #[cfg(test)]
 #[allow(clippy::approx_constant)]
 mod tests {
+    use crate::{array, assert_array_eq, builder::Builder, ops::is_nan};
     use float_eq::assert_float_eq;
-    use mlx_rs::{array, assert_array_eq, builder::Builder, ops::is_nan};
 
     use super::*;
