+++ mlx-rs/src/nn/embedding.rs
@@ -1,16 +1,17 @@
 //! Embedding layer.
 
+use crate::error::Exception;
+use crate::module::Module;
+use crate::module::Param;
+use crate::prelude::IndexOp;
+use crate::Array;
 use mlx_macros::ModuleParameters;
-use mlx_rs::error::Exception;
-use mlx_rs::module::Module;
-use mlx_rs::module::Param;
-use mlx_rs::prelude::IndexOp;
-use mlx_rs::Array;
 
 /// Implements a simple lookup table that maps each input integer to a high-dimensional vector.
 ///
 /// Typically used to embed discrete tokens for processing by neural networks.
 #[derive(Debug, Clone, ModuleParameters)]
+#[module(root = crate)]
 pub struct Embedding {
     /// The weight of the
     #[param]
@@ -27,8 +28,7 @@ impl Embedding {
     pub fn new(embedding_count: i32, dimensions: i32) -> Result<Self, Exception> {
         let scale = f32::sqrt(1.0 / (dimensions as f32));
         let weight =
-            mlx_rs::random::normal::<f32>(&[embedding_count, dimensions], None, None, None)?
-                * scale;
+            crate::random::normal::<f32>(&[embedding_count, dimensions], None, None, None)? * scale;
 
         Ok(Self {
             weight: Param::new(weight),
@@ -40,7 +40,7 @@ impl Embedding {
     /// Use this for example when input embedding and output projection
     /// weights are tied.
     pub fn as_linear(&self, x: &Array) -> Result<Array, Exception> {
-        mlx_rs::ops::matmul(x, self.weight.value.t())
+        crate::ops::matmul(x, self.weight.value.t())
     }
 }
 
@@ -63,10 +63,10 @@ mod tests {
 
     #[test]
     fn test_embedding() {
-        mlx_rs::random::seed(557).unwrap();
-        let a = mlx_rs::random::randint::<_, i32>(0, 10, &[2, 8, 8, 4], None).unwrap();
+        crate::random::seed(557).unwrap();
+        let a = crate::random::randint::<_, i32>(0, 10, &[2, 8, 8, 4], None).unwrap();
         assert_eq!(a.shape(), &[2, 8, 8, 4]);
-        assert_eq!(a.dtype(), mlx_rs::Dtype::Int32);
+        assert_eq!(a.dtype(), crate::Dtype::Int32);
         float_eq!(
             a.mean(None, None).unwrap().item::<f32>(),
             4.605_468_8,
@@ -80,7 +80,7 @@ mod tests {
 
         let result = Embedding::new(10, 8).unwrap().forward(&a).unwrap();
         assert_eq!(result.shape(), &[2, 8, 8, 4, 8]);
-        assert_eq!(result.dtype(), mlx_rs::Dtype::Float32);
+        assert_eq!(result.dtype(), crate::Dtype::Float32);
         float_eq!(
             result.mean(None, None).unwrap().item::<f32>(),
             -0.001_197_346_3,
