+++ mlx-rs/src/nn/mod.rs
@@ -5,11 +5,6 @@
 //! All modules provide a `new()` function that take mandatory parameters and other methods
 //! to set optional parameters.
 
-pub mod error;
-pub mod losses;
-pub mod macros;
-pub mod utils;
-
 mod activation;
 mod container;
 mod convolution;
@@ -24,7 +19,6 @@ mod quantized;
 mod recurrent;
 mod transformer;
 mod upsample;
-mod value_and_grad;
 
 pub use activation::*;
 pub use container::*;
@@ -40,9 +34,3 @@ pub use quantized::*;
 pub use recurrent::*;
 pub use transformer::*;
 pub use upsample::*;
-pub use value_and_grad::*;
-
-/// Re-export of the `mlx-nn-module` crate.
-pub mod module {
-    pub use mlx_rs::module::*;
-}
