+++ mlx-tests/tests/test_module.rs
@@ -1,7 +1,4 @@
-use mlx_macros::ModuleParameters;
-use mlx_nn::Linear;
-use mlx_rs::module::Module;
-use mlx_rs::{error::Exception, Array};
+use mlx_rs::{error::Exception, macros::ModuleParameters, module::Module, nn::Linear, Array};
 
 #[derive(Debug, ModuleParameters)]
 struct M {
