+++ mlx-tests/tests/test_module_parameters.rs
@@ -1,7 +1,9 @@
-use mlx_rs::module::{ModuleParameters, Param, Parameter};
-use mlx_rs::{array, Array};
-
-use mlx_nn::macros::ModuleParameters;
+use mlx_rs::{
+    array,
+    macros::ModuleParameters,
+    module::{ModuleParameters, Param, Parameter},
+    Array,
+};
 
 #[derive(ModuleParameters)]
 pub struct StructModule {
